Add mina gem
Instead of using capistrano, I have opted for using the mina gem for our
deployment scripts. Mina is very lightweight and does not require us to
configure or change our tasks to a mina-flavoured task. We can, but it's
not necessary. For now I am leaving the tasks as they are and
accomodating them in the deploy.rb file. When we settle on a deployment
tool/gem, we can flavour them to that tool's styling.